requirements (1).txt
streamlit
openai
pandas
python-dotenv
flask